Change name of OriginalFilename element
The name of the OriginalFilename element was changed to
OriginalFileLoadedIntoOmeka; this is a more accurate name,
since we are printing out the name of the file that was loaded
into Omeka. If we kept the old name OriginalFilename, it may be
misconstrued as the original picture file taken by the photographer
(usually a tiff), from which a JPEG is generated and loaded into
Omeka.
The function _buildOriginalFilenamesForItem was also renamed. The
new function name is _buildOriginalFilesLoadedIntoOmekaForItem