Add "item in context" information
This will be needed for the ingest into Fedora. Basically, we want the link to
an exhibition page that contains the item. Note that there are pseudo-exhibits,
such as the Omeka Demo exhibit, which may contain an item. So we check to make
sure that the exhibit name matches the Omeka collection name for the item.
Also, we just need one "item in context" value, so we just use the first valid
one we get.
This commit contains debugging code which needs to be removed.