Call CulCustomizePlugin::return_exhibit_pages_containing_current_item with argument
CulCustomizePlugin::return_exhibit_pages_containing_current_item now takes an argument
(commit 79e34fc0b51a5025888f5686c2a17260f9cb523d in CulCustomize)
This argument was needed to have the item-in-context functionality work in the
collection context - get_current_record('item') was not working, probably because
the current record is a collection. So cul-omeka-output was not working when used
in the Browse Items screen for a collection.